Restructures and renames code
