+++ frontend/app/src/components/accounts/exchanges/ExchangeAmountRow.vue
@@ -22,7 +22,7 @@
 import { BigNumber } from '@rotki/common';
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import { exchangeName } from '@/components/history/consts';
-import { SupportedExchange } from '@/services/balances/types';
+import { SupportedExchange } from '@/types/exchanges';
 
 @Component({
   name: 'ExchangeAmountRow'
