+++ frontend/app/src/components/history/consts.ts
@@ -4,26 +4,9 @@ import FOXIcon from '@/components/icons/FOXIcon.vue';
 import UpholdIcon from '@/components/icons/UpholdIcon.vue';
 import {
   EXCHANGE_BALANCER,
-  EXCHANGE_BINANCE,
-  EXCHANGE_BINANCEUS,
-  EXCHANGE_BITCOIN_DE,
-  EXCHANGE_BITFINEX,
-  EXCHANGE_BITMEX,
-  EXCHANGE_BITPANDA,
-  EXCHANGE_BITSTAMP,
-  EXCHANGE_BITTREX,
   EXCHANGE_BLOCKFI,
-  EXCHANGE_COINBASE,
-  EXCHANGE_COINBASEPRO,
   EXCHANGE_CRYPTOCOM,
-  EXCHANGE_FTX,
-  EXCHANGE_GEMINI,
-  EXCHANGE_ICONOMI,
-  EXCHANGE_INDEPENDENTRESERVE,
-  EXCHANGE_KRAKEN,
-  EXCHANGE_KUCOIN,
   EXCHANGE_NEXO,
-  EXCHANGE_POLONIEX,
   EXCHANGE_SHAPESHIFT,
   EXCHANGE_SUSHISWAP,
   EXCHANGE_UNISWAP,
@@ -37,6 +20,25 @@ import {
 } from '@/data/defaults';
 import i18n from '@/i18n';
 import { TradeLocation } from '@/services/history/types';
+import {
+  EXCHANGE_BINANCE,
+  EXCHANGE_BINANCEUS,
+  EXCHANGE_BITCOIN_DE,
+  EXCHANGE_BITFINEX,
+  EXCHANGE_BITMEX,
+  EXCHANGE_BITPANDA,
+  EXCHANGE_BITSTAMP,
+  EXCHANGE_BITTREX,
+  EXCHANGE_COINBASE,
+  EXCHANGE_COINBASEPRO,
+  EXCHANGE_FTX,
+  EXCHANGE_GEMINI,
+  EXCHANGE_ICONOMI,
+  EXCHANGE_INDEPENDENTRESERVE,
+  EXCHANGE_KRAKEN,
+  EXCHANGE_KUCOIN,
+  EXCHANGE_POLONIEX
+} from '@/types/exchanges';
 import { assert } from '@/utils/assertions';
 
 export const tradeLocations: TradeLocationData[] = [
