+++ frontend/app/src/components/settings/PremiumSettings.vue
@@ -111,7 +111,7 @@ import ConfirmDialog from '@/components/dialogs/ConfirmDialog.vue';
 import RevealableInput from '@/components/inputs/RevealableInput.vue';
 import { PremiumCredentialsPayload } from '@/store/session/types';
 import { ActionStatus } from '@/store/types';
-import { SettingsUpdate } from '@/typing/types';
+import { SettingsUpdate } from '@/types/user';
 import { trimOnPaste } from '@/utils/event';
 
 @Component({
