+++ frontend/app/src/components/settings/accounting/LedgerActionSettings.vue
@@ -34,8 +34,9 @@ import { Component, Mixins } from 'vue-property-decorator';
 import ActionStatusIndicator from '@/components/error/ActionStatusIndicator.vue';
 import SettingCategory from '@/components/settings/SettingCategory.vue';
 import SettingsMixin from '@/mixins/settings-mixin';
-import { ledgerActionsData, LedgerActionType } from '@/store/history/consts';
+import { ledgerActionsData } from '@/store/history/consts';
 import { ActionStatus } from '@/store/types';
+import { LedgerActionType } from '@/types/ledger-actions';
 
 const defaultTaxable: () => TaxableState = () => {
   const status: any = {};
