+++ frontend/app/src/components/settings/data-security/DataManagement.vue
@@ -43,11 +43,10 @@ import PurgeSelector, {
   PurgeParams
 } from '@/components/settings/data-security/PurgeSelector.vue';
 import StatusButton from '@/components/settings/data-security/StatusButton.vue';
-import { EXTERNAL_EXCHANGES, SUPPORTED_EXCHANGES } from '@/data/defaults';
-import { SupportedExchange } from '@/services/balances/types';
+import { EXTERNAL_EXCHANGES } from '@/data/defaults';
 import {
-  ALL_DECENTRALIZED_EXCHANGES,
   ALL_CENTRALIZED_EXCHANGES,
+  ALL_DECENTRALIZED_EXCHANGES,
   ALL_MODULES,
   ALL_TRANSACTIONS,
   Module
@@ -55,6 +54,7 @@ import {
 import { Purgeable } from '@/services/session/types';
 import { ACTION_PURGE_CACHED_DATA } from '@/store/session/const';
 import { ActionStatus } from '@/store/types';
+import { SUPPORTED_EXCHANGES, SupportedExchange } from '@/types/exchanges';
 
 @Component({
   components: { PurgeSelector, StatusButton, ConfirmDialog },
