+++ frontend/app/src/data/defaults.ts
@@ -1,3 +1,4 @@
+import { SUPPORTED_EXCHANGES } from '@/types/exchanges';
 import { CURRENCY_AFTER, CurrencyLocation } from '@/typing/types';
 
 export class Defaults {
@@ -16,24 +17,7 @@ export class Defaults {
   static DISPLAY_DATE_IN_LOCALTIME = true;
 }
 
-export const EXCHANGE_POLONIEX = 'poloniex';
-export const EXCHANGE_KRAKEN = 'kraken';
-export const EXCHANGE_BITTREX = 'bittrex';
-export const EXCHANGE_BITMEX = 'bitmex';
-export const EXCHANGE_BITPANDA = 'bitpanda';
-export const EXCHANGE_BITFINEX = 'bitfinex';
-export const EXCHANGE_BINANCE = 'binance';
-export const EXCHANGE_COINBASE = 'coinbase';
-export const EXCHANGE_COINBASEPRO = 'coinbasepro';
-export const EXCHANGE_GEMINI = 'gemini';
 export const EXCHANGE_CRYPTOCOM = 'cryptocom';
-export const EXCHANGE_BITSTAMP = 'bitstamp';
-export const EXCHANGE_BINANCEUS = 'binanceus';
-export const EXCHANGE_BITCOIN_DE = 'bitcoinde';
-export const EXCHANGE_ICONOMI = 'iconomi';
-export const EXCHANGE_KUCOIN = 'kucoin';
-export const EXCHANGE_FTX = 'ftx';
-export const EXCHANGE_INDEPENDENTRESERVE = 'independentreserve';
 export const EXCHANGE_BLOCKFI = 'blockfi';
 export const EXCHANGE_NEXO = 'nexo';
 export const EXCHANGE_SHAPESHIFT = 'shapeshift';
@@ -49,29 +33,6 @@ export const TRADE_LOCATION_EQUITIES = 'equities';
 export const TRADE_LOCATION_REALESTATE = 'realestate';
 export const TRADE_LOCATION_COMMODITIES = 'commodities';
 export const TRADE_LOCATION_BLOCKCHAIN = 'blockchain';
-
-/**
- * A list of exchanges that are supported via api connection
- */
-export const SUPPORTED_EXCHANGES = [
-  EXCHANGE_KRAKEN,
-  EXCHANGE_POLONIEX,
-  EXCHANGE_BITTREX,
-  EXCHANGE_BITMEX,
-  EXCHANGE_BITPANDA,
-  EXCHANGE_BITFINEX,
-  EXCHANGE_BINANCE,
-  EXCHANGE_BINANCEUS,
-  EXCHANGE_BITCOIN_DE,
-  EXCHANGE_COINBASE,
-  EXCHANGE_COINBASEPRO,
-  EXCHANGE_GEMINI,
-  EXCHANGE_ICONOMI,
-  EXCHANGE_BITSTAMP,
-  EXCHANGE_KUCOIN,
-  EXCHANGE_FTX,
-  EXCHANGE_INDEPENDENTRESERVE
-] as const;
 
 export const SUPPORTED_TRADE_LOCATIONS = [
   EXCHANGE_UNISWAP,
