+++ frontend/app/src/store/balances/mutations.ts
@@ -1,6 +1,5 @@
 import { SupportedAsset } from '@rotki/common/lib/data';
 import { MutationTree } from 'vuex';
-import { Exchange } from '@/model/action-result';
 import {
   Balances,
   BalanceType,
@@ -18,6 +17,7 @@ import {
   EditExchange,
   NonFungibleBalances
 } from '@/store/balances/types';
+import { Exchange } from '@/types/exchanges';
 import { ExchangeData, ExchangeInfo, ExchangeRates } from '@/typing/types';
 
 export const mutations: MutationTree<BalanceState> = {
