+++ frontend/app/src/types/ledger-actions.ts
@@ -1,0 +1,10 @@
+export enum LedgerActionType {
+  ACTION_INCOME = 'income',
+  ACTION_LOSS = 'loss',
+  ACTION_DONATION = 'donation received',
+  ACTION_EXPENSE = 'expense',
+  ACTION_DIVIDENDS = 'dividends income',
+  ACTION_AIRDROP = 'airdrop',
+  ACTION_GIFT = 'gift',
+  ACTION_GRANT = 'grant'
+}
