+++ frontend/app/src/views/ProfitLossReport.vue
@@ -103,7 +103,8 @@ import { Currency } from '@/model/currency';
 import { TaskType } from '@/model/task-type';
 import { ReportError, ReportPeriod } from '@/store/reports/types';
 import { Message } from '@/store/types';
-import { AccountingSettings, ProfitLossPeriod } from '@/typing/types';
+import { AccountingSettings } from '@/types/user';
+import { ProfitLossPeriod } from '@/typing/types';
 
 @Component({
   components: {
